+++ scratch.js
@@ -1,4 +1,4 @@
-import { again, test, reverse } from './src/index.js'
+import { learn, test, reverse, convert } from './src/index.js'
 // import pairs from '/Users/spencer/mountain/suffix-thumb/test/data/fr-words.js'
 // import pairs from '/Users/spencer/mountain/suffix-thumb/test/data/fr-nous.js'
 // import pairs from '/Users/spencer/mountain/suffix-thumb/test/data/future-simple.js'
@@ -13,7 +13,8 @@ let pairs = [
   ['ooocool', 'ooocool'],//unchanged
 ]
 
-let model = again(pairs)
+let model = learn(pairs)
+console.log(model)
 // console.log('----')
 // console.log(model)
 // console.log(reverse(model))
