+++ src/index.js
@@ -1,4 +1,4 @@
-import convert from './tiny/convert.js'
+import convert from './convert/index.js'
 // import compress from './compress/index.js'
 // import uncompress from './compress/uncompress.js'
 import reverse from './reverse/index.js'
@@ -8,11 +8,11 @@ import validate from './validate/index.js'
 // import classify from './classify/index.js'
 // import fingerprint from './fingerprint/index.js'
 
-import learn from './again/index.js'
+import learn from './learn/index.js'
 import compress from './pack/pack.js'
 import uncompress from './pack/unpack.js'
 import test from './test/index.js'
-import again from './again/index.js'
+// import again from './again/index.js'
 
 
-export { learn, convert, compress, uncompress, reverse, validate, test, again }
+export { learn, convert, compress, uncompress, reverse, validate, test }
