+++ src/learn/lib/getSuffix.js
