+++ src/learn/lib/goodEnough.js
